Remove constant from forkspec. Refactor LightClientBootstrapSchema
