+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/lightclient/LightClientBootstrap.java
@@ -15,50 +15,14 @@ package tech.pegasys.teku.spec.datastructures.lightclient;
 
 import tech.pegasys.teku.infrastructure.ssz.collections.SszBytes32Vector;
 import tech.pegasys.teku.infrastructure.ssz.containers.Container3;
-import tech.pegasys.teku.infrastructure.ssz.containers.ContainerSchema3;
-import tech.pegasys.teku.infrastructure.ssz.schema.collections.SszBytes32VectorSchema;
 import tech.pegasys.teku.infrastructure.ssz.tree.TreeNode;
-import tech.pegasys.teku.spec.config.SpecConfigAltair;
 import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlockHeader;
 import tech.pegasys.teku.spec.datastructures.state.SyncCommittee;
 
 public class LightClientBootstrap
     extends Container3<LightClientBootstrap, BeaconBlockHeader, SyncCommittee, SszBytes32Vector> {
 
-  public static class LightClientBootstrapSchema
-      extends ContainerSchema3<
-          LightClientBootstrap, BeaconBlockHeader, SyncCommittee, SszBytes32Vector> {
-
-    public LightClientBootstrapSchema(final SpecConfigAltair specConfigAltair) {
-      super(
-          "LightClientBootstrap",
-          namedSchema("header", BeaconBlockHeader.SSZ_SCHEMA),
-          namedSchema(
-              "current_sync_committee", new SyncCommittee.SyncCommitteeSchema(specConfigAltair)),
-          namedSchema(
-              "current_sync_committee_branch",
-              SszBytes32VectorSchema.create(specConfigAltair.getSyncCommitteeBranchLength())));
-    }
-
-    public LightClientBootstrap create(
-        BeaconBlockHeader beaconBlockHeader,
-        SyncCommittee syncCommittee,
-        SszBytes32Vector syncCommitteeBranch) {
-      return new LightClientBootstrap(this, beaconBlockHeader, syncCommittee, syncCommitteeBranch);
-    }
-
-    @SuppressWarnings("unchecked")
-    public SszBytes32VectorSchema<SszBytes32Vector> getSyncCommitteeBranchSchema() {
-      return (SszBytes32VectorSchema<SszBytes32Vector>) getChildSchema(2);
-    }
-
-    @Override
-    public LightClientBootstrap createFromBackingNode(TreeNode node) {
-      return new LightClientBootstrap(this, node);
-    }
-  }
-
-  protected LightClientBootstrap(
+  public LightClientBootstrap(
       final LightClientBootstrapSchema schema,
       final BeaconBlockHeader beaconBlockHeader,
       final SyncCommittee syncCommittee,
@@ -66,7 +30,8 @@ public class LightClientBootstrap
     super(schema, beaconBlockHeader, syncCommittee, syncCommitteeBranch);
   }
 
-  private LightClientBootstrap(final LightClientBootstrapSchema type, final TreeNode backingNode) {
+  protected LightClientBootstrap(
+      final LightClientBootstrapSchema type, final TreeNode backingNode) {
     super(type, backingNode);
   }
 }
