+++ ethereum/spec/src/test/resources/tech/pegasys/teku/spec/config/altairForkSpec.json
@@ -74,7 +74,6 @@
   "SLOTS_PER_EPOCH": "32",
   "SLOTS_PER_HISTORICAL_ROOT": "8192",
   "SYNC_COMMITTEE_SIZE": "512",
-  "SYNC_COMMITTEE_BRANCH_LENGTH": "5",
   "SYNC_COMMITTEE_SUBNET_COUNT": "4",
   "TARGET_AGGREGATORS_PER_COMMITTEE": "16",
   "TARGET_AGGREGATORS_PER_SYNC_SUBCOMMITTEE": "16",
